refactorings + code cleanups
